Define special types model in the sandbox
